feat(replays): Inject an `init-replay` breadcrumb so there is always one in the list
